+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * refactor: spin off `DomainEventStore` from `DomainEventAwareEntityManager` for
   better separation of concerns & easy reuse.
 * feat: full transaction support for postFlush
+* fix: override `transactional()` and `wrapInTransaction()`
 
 ## 2.1.0